Refactor to use macros 1.1.
It's slightly incomplete still but easier than before.
