+++ src/lib.rs
@@ -1,10 +1,11 @@
-#![feature(associated_consts, custom_derive, custom_attribute)]
+#![feature(associated_consts, proc_macro)]
 
 extern crate byteorder;
 extern crate chrono;
 extern crate crypto;
 extern crate crc;
 extern crate num;
+#[macro_use] extern crate tl_macros;
 
 pub mod tl;
 pub mod rpc;
